+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/certifiedProduct/validation/CertifiedProductValidatorImpl.java
@@ -14,6 +14,8 @@ import gov.healthit.chpl.dao.CertifiedProductDAO;
 import gov.healthit.chpl.dao.DeveloperDAO;
 import gov.healthit.chpl.dao.EntityRetrievalException;
 import gov.healthit.chpl.dao.TestingLabDAO;
+import gov.healthit.chpl.domain.CertificationResult;
+import gov.healthit.chpl.domain.CertificationResultAdditionalSoftware;
 import gov.healthit.chpl.domain.CertifiedProductSearchDetails;
 import gov.healthit.chpl.dto.CertificationBodyDTO;
 import gov.healthit.chpl.dto.CertificationEditionDTO;
@@ -24,10 +26,12 @@ import gov.healthit.chpl.dto.PendingCertificationResultAdditionalSoftwareDTO;
 import gov.healthit.chpl.dto.PendingCertificationResultDTO;
 import gov.healthit.chpl.dto.PendingCertifiedProductDTO;
 import gov.healthit.chpl.dto.TestingLabDTO;
+import gov.healthit.chpl.manager.CertifiedProductManager;
 import gov.healthit.chpl.util.CertificationResultRules;
 
 public class CertifiedProductValidatorImpl implements CertifiedProductValidator {
 	@Autowired CertifiedProductDAO cpDao;
+	@Autowired CertifiedProductManager cpManager;
 	@Autowired TestingLabDAO atlDao;
 	@Autowired CertificationEditionDAO certEditionDao;
 	@Autowired CertificationBodyDAO acbDao;
@@ -177,8 +181,73 @@ public class CertifiedProductValidatorImpl implements CertifiedProductValidator 
 		}
 	}
 	@Override
-	public void validate(CertifiedProductSearchDetails product) {
-		//TODO: not sure if we should do the same validation here or not
+	public void validate(CertifiedProductSearchDetails product) {		
+		//if it's a new product, check the id parts
+		String uniqueId = product.getChplProductNumber();
+		String[] uniqueIdParts = uniqueId.split("\\.");
+		if(uniqueIdParts != null && uniqueIdParts.length == 9) {
+			
+			//validate that these pieces match up with data
+			String icsCode = uniqueIdParts[6];
+			String additionalSoftwareCode = uniqueIdParts[7];
+			String certifiedDateCode = uniqueIdParts[8];
+			
+			if(icsCode.equals("0") && product.getIcs().equals(Boolean.TRUE)) {
+				product.getErrorMessages().add("The unique id indicates the product does not have ICS but the value for Inherited Certification Status is true.");
+			} else if(!icsCode.equals("0") && product.getIcs().equals(Boolean.FALSE)) {
+				product.getErrorMessages().add("The unique id indicates the product does have ICS but the value for Inherited Certification Status is false.");
+			}
+			if(additionalSoftwareCode.equals("0")) {
+				boolean hasAS = false;
+				for(CertificationResult cert : product.getCertificationResults()) {
+					if(cert.getAdditionalSoftware() != null && cert.getAdditionalSoftware().size() > 0) {
+						hasAS = true;
+					}
+				}
+				if(hasAS) {
+					product.getErrorMessages().add("The unique id indicates the product does not have additional software but some is listed for the product.");
+				}
+			} else if(additionalSoftwareCode.equals("1")) {
+				boolean hasAS = false;
+				for(CertificationResult cert : product.getCertificationResults()) {
+					if(cert.getAdditionalSoftware() != null && cert.getAdditionalSoftware().size() > 0) {
+						hasAS = true;
+					}
+				}
+				if(!hasAS) {
+					product.getErrorMessages().add("The unique id indicates the product has additional software but none is listed for the product.");
+				}
+			} else {
+				product.getErrorMessages().add("The additional software part of the unique ID must be 0 or 1.");
+			}
+			SimpleDateFormat idDateFormat = new SimpleDateFormat("yyMMdd");
+			try {
+				Date idDate = idDateFormat.parse(certifiedDateCode);
+				if(product.getCertificationDate() == null || 
+						idDate.getTime() != product.getCertificationDate().longValue()) {
+					product.getErrorMessages().add("The certification date provided in the unique id does not match the certification date of the product.");
+				}
+			} catch (ParseException pex) {
+				product.getErrorMessages().add("Could not parse the certification date part of the product id: " + certifiedDateCode);
+			}
+		}
+		
+		validateDemographics(product);
+		
+		for(CertificationResult cert : product.getCertificationResults()) {
+			if(cert.getAdditionalSoftware() != null && cert.getAdditionalSoftware().size() > 0) {
+				for(CertificationResultAdditionalSoftware asDto : cert.getAdditionalSoftware()) {
+					if(asDto.getCertifiedProductId() == null && !StringUtils.isEmpty(asDto.getCertifiedProductNumber())) {
+						try {
+							boolean exists = cpManager.chplIdExists(asDto.getCertifiedProductNumber());
+							if(!exists) {
+								product.getErrorMessages().add("No CHPL product was found matching additional software " + asDto.getCertifiedProductNumber() + " for " + cert.getNumber());
+							}
+						} catch(Exception ex) {}
+					}
+				}
+			}
+		}
 	}
 	
 	protected void validateDemographics(PendingCertifiedProductDTO product) {
@@ -267,10 +336,10 @@ public class CertifiedProductValidatorImpl implements CertifiedProductValidator 
 			product.getErrorMessages().add("ICS is required.");
 		}
 		
-		if(!StringUtils.isEmpty(product.getTransparencyAttestationUrl()) && 
-				urlRegex.matcher(product.getTransparencyAttestationUrl()).matches() == false) {
-			product.getErrorMessages().add("Transparency attestation URL is not a valid URL format.");
-		}
+//		if(!StringUtils.isEmpty(product.getTransparencyAttestationUrl()) && 
+//				urlRegex.matcher(product.getTransparencyAttestationUrl()).matches() == false) {
+//			product.getErrorMessages().add("Transparency attestation URL is not a valid URL format.");
+//		}
 		
 		for(PendingCertificationResultDTO cert : product.getCertificationCriterion()) {
 			if(cert.getMeetsCriteria() == null) {
@@ -284,6 +353,60 @@ public class CertifiedProductValidatorImpl implements CertifiedProductValidator 
 						(cert.getTestProcedures() == null || cert.getTestProcedures().size() == 0)) {
 					product.getErrorMessages().add("Test Procedures are required for certification " + cert.getNumber() + ".");
 				}
+			}
+		}
+	}
+	
+	protected void validateDemographics(CertifiedProductSearchDetails product) {
+		if(product.getCertificationEdition() == null || product.getCertificationEdition().get("id") == null) {
+			product.getErrorMessages().add("Certification edition is required but was not found.");
+		}
+		if(StringUtils.isEmpty(product.getAcbCertificationId())) {
+			product.getErrorMessages().add("CHPL certification ID is required but was not found.");
+		}
+		if(product.getCertificationDate() == null) {
+			product.getErrorMessages().add("Certification date was not found.");
+		} else if(product.getCertificationDate() > new Date().getTime()) {
+			product.getErrorMessages().add("Certification date occurs in the future.");
+		}
+		if(product.getCertifyingBody() == null || product.getCertifyingBody().get("id") == null) {
+			product.getErrorMessages().add("ACB ID is required but was not found.");
+		}
+		
+		if(product.getDeveloper() == null) {
+			product.getErrorMessages().add("A developer is required.");
+		}
+		
+		if(product.getProduct() == null || StringUtils.isEmpty(product.getProduct().get("name"))) {
+			product.getErrorMessages().add("A product name is required.");
+		}
+		
+		if(product.getProduct() == null || StringUtils.isEmpty(product.getProduct().get("version"))) {
+			product.getErrorMessages().add("A product version is required.");
+		}
+
+		if(product.getIcs() == null) {
+			product.getErrorMessages().add("ICS is required.");
+		}
+		
+//		if(!StringUtils.isEmpty(product.getTransparencyAttestationUrl()) && 
+//				urlRegex.matcher(product.getTransparencyAttestationUrl()).matches() == false) {
+//			product.getErrorMessages().add("Transparency attestation URL is not a valid URL format.");
+//		}
+		
+		for(CertificationResult cert : product.getCertificationResults()) {
+			if(cert.isSuccess() != null && cert.isSuccess() == Boolean.TRUE) {
+				if(certRules.hasCertOption(cert.getNumber(), CertificationResultRules.GAP) &&
+						cert.isGap() == null) {
+					product.getErrorMessages().add("GAP is required for certification " + cert.getNumber() + ".");
+				}
+				//Jennifer asked to take out the test procedure validation for existing products
+				//so that when users are on the edit screen, they are not required
+				//to have test procedures for all certifications
+//				if(certRules.hasCertOption(cert.getNumber(), CertificationResultRules.TEST_PROCEDURE_VERSION) &&
+//						(cert.getTestProcedures() == null || cert.getTestProcedures().size() == 0)) {
+//					product.getErrorMessages().add("Test Procedures are required for certification " + cert.getNumber() + ".");
+//				}
 			}
 		}
 	}