+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/dao/CertifiedProductAccessibilityStandardDAO.java
@@ -7,6 +7,7 @@ import java.util.List;
 public interface CertifiedProductAccessibilityStandardDAO {
 	
 	public List<CertifiedProductAccessibilityStandardDTO> getAccessibilityStandardsByCertifiedProductId(Long certifiedProductId) throws EntityRetrievalException;
+	public CertifiedProductAccessibilityStandardDTO lookupMapping(Long certifiedProductId, Long accStdId) throws EntityRetrievalException; 
 	public CertifiedProductAccessibilityStandardDTO createCertifiedProductAccessibilityStandard(CertifiedProductAccessibilityStandardDTO toCreate) throws EntityCreationException;
 	public CertifiedProductAccessibilityStandardDTO deleteCertifiedProductAccessibilityStandards(Long id) throws EntityRetrievalException;